pyproject.toml
[project]
name = "ai-startups"
version = "0.1.0"
description = "Daily-updated open-data jobboard covering the EU AI/tech ecosystem."
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MIT" }
authors = [{ name = "Aramente" }]
dependencies = [
    "httpx>=0.27",
    "pydantic>=2.7",
    "pyarrow>=17.0",
    "click>=8.1",
    "pyyaml>=6.0",
    "tenacity>=9.0",
    "python-dateutil>=2.9",
    "markdownify>=0.13",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3",
    "pytest-asyncio>=0.24",
    "respx>=0.21",
    "ruff>=0.7",
]

[project.scripts]
pipeline = "pipeline.cli:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["pipeline"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "SIM"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
markers = [
    "live: tests that hit real network endpoints (skip in CI by default)",
]
testpaths = ["tests"]
